Move set metadata.
